+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Changing the values for Registration helper text and Account confirmation in Admin > Settings > Registration doesn't cause other values to be erased anymore.
 
+### Added
+
+- Configure sign-up button (custom link) on homepage banner
+
 ## 2022-01-05
 
 ### Changed